First Function
I spent some time today taking  some first steps of getting a student profile which WRITES data from a student form (based on Assignment I question III) to a table in the 'soen287' database.

In order to test this code for yourself, you will have needed to follow Nicolas's DB creation instructions:
https://github.com/n-aubry/soen287-project/pull/8

Open this database in your localhost, and import the students.sql table to create the table instance.

I have only created a function for writing to the database, I will be implementing READ and EDIT in the following days.