Updated University .json files
I added minimum gpa, minimum SAT for each program for each school. I also added the average entrance GPA for each school, average SAT and admission rates for each school.
I then added a few scholarships and bursaries for each school, which each have requirements (gpa, sat, financial need + the value of the aid).

The calculation for financial need is :
Tuition - [(Household income - 60,000)*0.15].

The value of the scholarships is the maximum amount receivable, and the indicated financial need is the minimum.