MultiStep Form
Beginnings of multiStep form to be responsible for studentProfileCreation. Currently uncomplete.

I) Need to add steps for desired attributes of other universities. (plan to include photos of campuses, with brief description of universities for examples)

II) Need to autoupdate dropdown menu with list of possible majors

III) Need to decide on which financial information will be relevant for match