updated StudentProfileCreation
This is the isolated updated for Student profile creation form from my other messy PR.

The only thing I've done that I'm aware might need to be changed is hardcoded the preferredRanking. I wasn't sure what was happening with that field in terms of the Match-Me, so it is currently a hidden input field.